+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Upgrade dependencies
 + Add generic detail
 + Add ModalWindow component
++ Add transaction detail
 
 ## 0.9.0
 + Upgrade dependencies